backend/requirements.txt
alembic==1.14.1
aniso8601==10.0.0
bcrypt==4.2.1
blinker==1.9.0
cachelib==0.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Caching==2.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
flask-talisman==1.1.0
Flask-WTF==1.2.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.9
MarkupSafe==3.0.2
marshmallow==3.26.1
packaging==24.2
Pillow==10.0.0
PyJWT==2.10.1
python-dotenv==1.0.0
python-slugify==8.0.4
pytz==2025.1
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.38
stripe==8.0.0
text-unidecode==1.3
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
WTForms==3.2.1
